cloud_deployment/functions/extract_website_data/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.21.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 